Move annotations to separate files
