--- tests/storage/annotations/OpenApi/Anotations.php
@@ -1,142 +1,0 @@
-<?php
-
-/**
- * @OA\Info(
- *      version="1.0.0",
- *      title="L5 OpenApi",
- *      description="L5 Swagger OpenApi description",
- *      @OA\Contact(
- *          email="darius@matulionis.lt"
- *      ),
- *     @OA\License(
- *         name="Apache 2.0",
- *         url="http://www.apache.org/licenses/LICENSE-2.0.html"
- *     )
- * )
- */
-
-class L5SwaggerAnnotationsExampleInfo {}
-
-/**
- *  @OA\Server(
- *      url=L5_SWAGGER_CONST_HOST,
- *      description="L5 Swagger OpenApi dynamic host server"
- *  )
- *
- *  @OA\Server(
-*      url="https://projects.dev/api/v1",
- *      description="L5 Swagger OpenApi Server"
- * )
- */
-class L5SwaggerAnnotationsExampleServer {}
-
-/**
- * @OA\SecurityScheme(
- *     type="oauth2",
- *     description="Use a global client_id / client_secret and your username / password combo to obtain a token",
- *     name="Password Based",
- *     in="header",
- *     scheme="https",
- *     securityScheme="Password Based",
- *     @OA\Flow(
- *         flow="password",
- *         authorizationUrl="/oauth/authorize",
- *         tokenUrl="/oauth/token",
- *         refreshUrl="/oauth/token/refresh",
- *         scopes={}
- *     )
- * )
- */
-class L5SwaggerAnnotationsExampleSecurityScheme {}
-
-/**
- * @OA\OpenApi(
- *   security={
- *     {
- *       "oauth2": {"read:oauth2"},
- *     }
- *   }
- * )
- */
-class L5SwaggerAnnotationsExampleSecurity {}
-
-/**
- * @OA\Tag(
- *     name="project",
- *     description="Everything about your Projects",
- *     @OA\ExternalDocumentation(
- *         description="Find out more",
- *         url="http://swagger.io"
- *     )
- * )
- *
- * @OA\Tag(
- *     name="user",
- *     description="Operations about user",
- *     @OA\ExternalDocumentation(
- *         description="Find out more about",
- *         url="http://swagger.io"
- *     )
- * )
- * @OA\ExternalDocumentation(
- *     description="Find out more about Swagger",
- *     url="http://swagger.io"
- * )
- */
-class L5SwaggerAnnotationsExampleTag {}
-
-
-class L5SwaggerAnnotationsExampleProjects
-{
-    /**
-     * @OA\Get(
-     *      path="/projects",
-     *      operationId="getProjectsList",
-     *      tags={"Projects"},
-     *      summary="Get list of projects",
-     *      description="Returns list of projects",
-     *      @OA\Response(
-     *          response=200,
-     *          description="successful operation"
-     *       ),
-     *       @OA\Response(response=400, description="Bad request"),
-     *       security={
-     *           {"api_key_security_example": {}}
-     *       }
-     *     )
-     *
-     * Returns list of projects
-     */
-    public function getProjectsList() {}
-
-    /**
-     * @OA\Get(
-     *      path="/projects/{id}",
-     *      operationId="getProjectById",
-     *      tags={"Projects"},
-     *      summary="Get project information",
-     *      description="Returns project data",
-     *      @OA\Parameter(
-     *          name="id",
-     *          description="Project id",
-     *          required=true,
-     *          in="path",
-     *          @OA\Schema(
-     *              type="integer"
-     *          )
-     *      ),
-     *      @OA\Response(
-     *          response=200,
-     *          description="successful operation"
-     *       ),
-     *      @OA\Response(response=400, description="Bad request"),
-     *      @OA\Response(response=404, description="Resource Not Found"),
-     *      security={
-     *         {
-     *             "oauth2_security_example": {"write:projects", "read:projects"}
-     *         }
-     *     },
-     * )
-     */
-    public function getProjectById() {}
-}
