+++ tests/storage/annotations/OpenApi/L5SwaggerAnnotationsExampleInfo.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace Tests\storage\annotations\OpenApi;
+
+/**
+ * @OA\Info(
+ *      version="1.0.0",
+ *      x={
+ *          "logo": {
+ *              "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
+ *          }
+ *      },
+ *      title="L5 OpenApi",
+ *      description="L5 Swagger OpenApi description",
+ *      @OA\Contact(
+ *          email="darius@matulionis.lt"
+ *      ),
+ *     @OA\License(
+ *         name="Apache 2.0",
+ *         url="https://www.apache.org/licenses/LICENSE-2.0.html"
+ *     )
+ * )
+ */
+
+class L5SwaggerAnnotationsExampleInfo {}
