+++ tests/storage/annotations/OpenApi/L5SwaggerAnnotationsExampleProjects.php
@@ -1,0 +1,58 @@
+<?php
+
+namespace Tests\storage\annotations\OpenApi;
+
+class L5SwaggerAnnotationsExampleProjects
+{
+    /**
+     * @OA\Get(
+     *      path="/projects",
+     *      operationId="getProjectsList",
+     *      tags={"Projects"},
+     *      summary="Get list of projects",
+     *      description="Returns list of projects",
+     *      @OA\Response(
+     *          response=200,
+     *          description="successful operation"
+     *       ),
+     *       @OA\Response(response=400, description="Bad request"),
+     *       security={
+     *           {"api_key_security_example": {}}
+     *       }
+     *     )
+     *
+     * Returns list of projects
+     */
+    public function getProjectsList() {}
+
+    /**
+     * @OA\Get(
+     *      path="/projects/{id}",
+     *      operationId="getProjectById",
+     *      tags={"Projects"},
+     *      summary="Get project information",
+     *      description="Returns project data",
+     *      @OA\Parameter(
+     *          name="id",
+     *          description="Project id",
+     *          required=true,
+     *          in="path",
+     *          @OA\Schema(
+     *              type="integer"
+     *          )
+     *      ),
+     *      @OA\Response(
+     *          response=200,
+     *          description="successful operation"
+     *       ),
+     *      @OA\Response(response=400, description="Bad request"),
+     *      @OA\Response(response=404, description="Resource Not Found"),
+     *      security={
+     *         {
+     *             "oauth2_security_example": {"write:projects", "read:projects"}
+     *         }
+     *     },
+     * )
+     */
+    public function getProjectById() {}
+}
