+++ tests/storage/annotations/OpenApi/L5SwaggerAnnotationsExampleSecurity.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Tests\storage\annotations\OpenApi;
+
+/**
+ * @OA\OpenApi(
+ *   security={
+ *     {
+ *       "oauth2": {"read:oauth2"},
+ *     }
+ *   }
+ * )
+ */
+class L5SwaggerAnnotationsExampleSecurity {}
