+++ tests/storage/annotations/OpenApi/L5SwaggerAnnotationsExampleSecurityScheme.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace Tests\storage\annotations\OpenApi;
+
+/**
+ * @OA\SecurityScheme(
+ *     type="oauth2",
+ *     description="Use a global client_id / client_secret and your username / password combo to obtain a token",
+ *     name="Password Based",
+ *     in="header",
+ *     scheme="https",
+ *     securityScheme="Password Based",
+ *     @OA\Flow(
+ *         flow="password",
+ *         authorizationUrl="/oauth/authorize",
+ *         tokenUrl="/oauth/token",
+ *         refreshUrl="/oauth/token/refresh",
+ *         scopes={}
+ *     )
+ * )
+ */
+class L5SwaggerAnnotationsExampleSecurityScheme {}
