+++ tests/storage/annotations/OpenApi/L5SwaggerAnnotationsExampleServer.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Tests\storage\annotations\OpenApi;
+
+/**
+ *  @OA\Server(
+ *      url=L5_SWAGGER_CONST_HOST,
+ *      description="L5 Swagger OpenApi dynamic host server"
+ *  )
+ *
+ *  @OA\Server(
+ *      url="https://projects.dev/api/v1",
+ *      description="L5 Swagger OpenApi Server"
+ * )
+ */
+class L5SwaggerAnnotationsExampleServer {}
